requirements/local.txt
-r base.txt
django-debug-toolbar
ipdb
ipython
